+++ client/app/hearingSchedule/components/AssignHearings.jsx
@@ -1,31 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import _ from 'lodash';
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
 import Link from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/Link';
 import COPY from '../../../COPY.json';
-import SearchableDropdown from '../../components/SearchableDropdown';
+import RoSelectorDropdown from './RoSelectorDropdown';
 
 export default class AssignHearings extends React.Component {
-
-  regionalOfficeOptions = () => {
-    let regionalOfficeDropdowns = [];
-
-    _.forEach(this.props.regionalOffices, (value, key) => {
-      regionalOfficeDropdowns.push({
-        label: `${value.city}, ${value.state}`,
-        value: key
-      });
-    });
-
-    regionalOfficeDropdowns.push({
-      label: 'Central Office',
-      value: 'CO'
-    });
-
-    return _.orderBy(regionalOfficeDropdowns, (ro) => ro.label, 'asc');
-  };
-
   render() {
     return <AppSegment filledBackground>
       <h1>{COPY.HEARING_SCHEDULE_ASSIGN_HEARINGS_HEADER}</h1>
@@ -34,13 +14,10 @@ export default class AssignHearings extends React.Component {
         to="/schedule">
         {COPY.HEARING_SCHEDULE_ASSIGN_HEARINGS_VIEW_SCHEDULE_LINK}
       </Link>
-      <SearchableDropdown
-        name="ro"
-        label="Regional Office"
-        options={this.regionalOfficeOptions()}
+      <RoSelectorDropdown
+        regionalOffices={this.props.regionalOffices}
         onChange={this.props.onRegionalOfficeChange}
         value={this.props.selectedRegionalOffice}
-        placeholder=""
       />
     </AppSegment>;
   }
